media: ti-vpe: cal: combine wdma irq handling
Instead of handling the WDMA START and END interrupts separately, we
need to handle both at the same time to better manage the inherent race
conditions related to CAL interrupts.

Change the code so that we have a single function,
cal_irq_handle_wdma(), which gets two booleans, start and end, as
parameters, which allows us to manage the race conditions in following
patches.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>