media: imx219: Switch from open to init_cfg
Use the init_cfg pad level operation instead of the internal subdev
open operation to set default formats on the pads. Add a dedicated
imx219_init_formats() functions for this, as it will be called later
when routing support will be added.
While at it, make the imx219_pad_ops more easier to read.

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@ideasonboard.com>